+++ blocks/rss-feature-block/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wpmedia/sitemaps-feature-block
 
+## 0.1.0-tk.1
+
+### Minor Changes
+
+- Added rss to prop-types
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wpmedia/feeds-content-elements@0.1.0-tk.2
+  - @wpmedia/feeds-prop-types@0.2.0-tk.0
+
 ## 0.1.0-tk.0
 
 ### Minor Changes