+++ utils/prop-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wpmedia/feeds-prop-types
 
+## 0.2.0-tk.0
+
+### Minor Changes
+
+- Added rss to prop-types
+
 ## 0.1.1
 
 ### Patch Changes